refactor: voting naming and tests
