+++ frontend/src/lib/utils/proposals.utils.ts
@@ -13,7 +13,7 @@ import { get } from "svelte/store";
 import { PROPOSAL_COLOR } from "../constants/proposals.constants";
 import { i18n } from "../stores/i18n";
 import type { ProposalsFiltersStore } from "../stores/proposals.store";
-import type { VoteInProgress } from "../stores/voting.store";
+import type { VoteRegistration } from "../stores/vote-registration.store";
 import type { Color } from "../types/theme";
 import { nowInSeconds } from "./date.utils";
 import { errorToString } from "./error.utils";
@@ -505,7 +505,7 @@ export const registerVoteErrorDetails = ({
 };
 
 /** There are neurons in a queue whose vote is not yet been registered */
-export const voteRegistrationActive = (votes: VoteInProgress[]): boolean =>
+export const voteRegistrationActive = (votes: VoteRegistration[]): boolean =>
   votes.find(
     ({ neuronIds, successfullyVotedNeuronIds }) =>
       neuronIds.length > successfullyVotedNeuronIds.length
