+++ frontend/src/tests/lib/components/ContextWrapperTest.svelte
@@ -4,8 +4,9 @@
   export let Component: SvelteComponent;
   export let contextKey: string;
   export let contextValue: unknown;
+  export let props: object = {};
 
   setContext(contextKey, contextValue);
 </script>
 
-<svelte:component this={Component} on:nnsClose />
+<svelte:component this={Component} {...props} on:nnsClose />
