+++ frontend/src/tests/lib/components/proposal-detail/VotingCard/VotingCard.spec.ts
@@ -7,15 +7,22 @@ import { GovernanceCanister, ProposalStatus, Vote } from "@dfinity/nns";
 import { fireEvent, screen } from "@testing-library/dom";
 import { render, waitFor } from "@testing-library/svelte";
 import { tick } from "svelte";
+import { writable } from "svelte/store";
+import VotingCard from "../../../../../lib/components/proposal-detail/VotingCard/VotingCard.svelte";
 import { SECONDS_IN_YEAR } from "../../../../../lib/constants/constants";
 import { authStore } from "../../../../../lib/stores/auth.store";
 import { neuronsStore } from "../../../../../lib/stores/neurons.store";
 import { votingNeuronSelectStore } from "../../../../../lib/stores/proposals.store";
+import {
+  SELECTED_PROPOSAL_CONTEXT_KEY,
+  type SelectedProposalContext,
+  type SelectedProposalStore,
+} from "../../../../../lib/types/selected-proposal.context";
 import { mockAuthStoreSubscribe } from "../../../../mocks/auth.store.mock";
 import { MockGovernanceCanister } from "../../../../mocks/governance.canister.mock";
 import { mockNeuron } from "../../../../mocks/neurons.mock";
 import { mockProposalInfo } from "../../../../mocks/proposal.mock";
-import VotingCardTest from "./VotingCardTest.svelte";
+import ContextWrapperTest from "../../ContextWrapperTest.svelte";
 
 jest.mock("../../../../../lib/utils/html.utils", () => ({
   sanitizeHTML: (value) => Promise.resolve(value),
@@ -37,9 +44,19 @@ describe("VotingCard", () => {
   }));
 
   const renderVotingCard = () =>
-    render(VotingCardTest, {
+    render(ContextWrapperTest, {
       props: {
-        proposalInfo,
+        props: {
+          proposalInfo,
+        },
+        contextKey: SELECTED_PROPOSAL_CONTEXT_KEY,
+        contextValue: {
+          store: writable<SelectedProposalStore>({
+            proposalId: proposalInfo.id,
+            proposal: proposalInfo,
+          }),
+        } as SelectedProposalContext,
+        Component: VotingCard,
       },
     });
 
@@ -97,7 +114,10 @@ describe("VotingCard", () => {
 
       jest
         .spyOn(GovernanceCanister, "create")
-        .mockImplementation((): GovernanceCanister => mockGovernanceCanister);
+        .mockImplementation(
+          (): GovernanceCanister =>
+            mockGovernanceCanister as unknown as GovernanceCanister
+        );
       spyRegisterVote = jest.spyOn(mockGovernanceCanister, "registerVote");
       spyListNeurons = jest.spyOn(mockGovernanceCanister, "listNeurons");
 
@@ -127,12 +147,12 @@ describe("VotingCard", () => {
     });
 
     // it's on to show "console.error('vote:..." in the output (because of NO mock in canister)
-    it("should trigger register-vote NO", async () => {
+    it.only("should trigger register-vote NO", async () => {
       await fireEvent.click(screen.queryByTestId("vote-no") as Element);
       await fireEvent.click(screen.queryByTestId("confirm-yes") as Element);
       await waitFor(() =>
-        expect(spyRegisterVote).toBeCalledWith({
-          neuronId: neuronIds[0],
+        expect(spyRegisterVote).toHaveBeenCalledWith({
+          neuronId: neuronIds[1],
           vote: Vote.NO,
           proposalId: proposalInfo.id,
         })
